Move geolocation rules to base of `data` dir [#6]
Previously, the workflow was downloading the same file into the
strain-specific sub-dirs of `data` — even though the data was exactly
the same in all cases — resulting in 4 identical copies of the
`general_geolocation_rules` output file.

Now we download the upstream input once, concat the local addition
once, and end up with one copy of the output, which we use each time
in the subsequent `curate` rule.