Refactor config to allow override of species list [#8]
[Slack
context](https://bedfordlab.slack.com/archives/C0K3GS3J8/p1716315967004439)
with discussion of different ways to accomplish this.

I think overall this approach is better; it does cause an extremely
tight coupling between the two parts of the config, but that coupling
was already present in the old `VIRUSES` global in the Snakemake
files; this mainly changes where that list lives.